:recycle: Lib dir for GrqphQL queries
no more svg on images folder ans separated hooks with libs
