+++ jsconfig.json
@@ -14,6 +14,9 @@
             "@hooks/*": [
                 "src/hooks/*"
             ],
+            "@lib/*": [
+                "src/lib/*"
+            ],
             "@images/*": [
                 "static/images/*"
             ],
