+++ src/@portfolio-ui/Icons/moon.js
@@ -2,21 +2,7 @@ import React from 'react';
 
 const Moon = () => {
   return (
-    <svg
-      version="1.1"
-      id="Layer_1"
-      xmlns="http://www.w3.org/2000/svg"
-      x="0px"
-      y="0px"
-      viewBox="0 0 100 92.3"
-      style="enable-background:new 0 0 100 92.3;"
-    >
-      <path
-        class="st0"
-        d="M59.7,21.3c-3.8-1.6-8-2.4-12.2-2.1c10.9,5.6,15.2,18.9,9.6,29.7S38.2,64,27.4,58.5c-1.2-0.6-2.4-1.4-3.5-2.2
-   c5.6,13.9,21.4,20.6,35.2,15C73,65.7,79.7,49.9,74.1,36C71.5,29.4,66.3,24.1,59.7,21.3z"
-      />
-    </svg>
+    <svg viewBox="0 0 24 24" width="30" height="30" fill="#FFC024"><path d="M15.977 14.456c-3.839 0-6.294-2.393-6.294-6.217 0-.79.192-1.92.447-2.505a.9.9 0 00.078-.332A.401.401 0 009.79 5c-.078 0-.249.021-.405.078C6.76 6.122 5 8.93 5 11.888 5 16.035 8.179 19 12.337 19c3.058 0 5.705-1.842 6.585-4.142.064-.162.078-.332.078-.395a.437.437 0 00-.419-.438.98.98 0 00-.305.064 7.979 7.979 0 01-2.299.367z"></path></svg>
   );
 };
