+++ src/@portfolio-ui/Icons/sun.js
@@ -2,69 +2,8 @@ import React from 'react';
 
 const Sun = () => {
   return (
-    <svg
-      version="1.1"
-      xmlns="http://www.w3.org/2000/svg"
-      x="0px"
-      y="0px"
-      viewBox="0 0 100 92.3"
-      style="enable-background:new 0 0 100 92.3;"
-    >
-      <g id="Layer_1" class="st0">
-        <path
-          class="st1"
-          d="M59.7,21.3c-3.8-1.6-8-2.4-12.2-2.1c10.9,5.6,15.2,18.9,9.6,29.7S38.2,64,27.4,58.5c-1.2-0.6-2.4-1.4-3.5-2.2
-     c5.6,13.9,21.4,20.6,35.2,15C73,65.7,79.7,49.9,74.1,36C71.5,29.4,66.3,24.1,59.7,21.3z"
-        />
-      </g>
-      <g id="Layer_2">
-        <g>
-          <path
-            class="st2"
-            d="M50,65c-10.5,0-18.8-8.3-18.8-18.8S39.5,27.4,50,27.4s18.8,8.3,18.8,18.8S60.5,65,50,65z"
-          />
-          <path
-            class="st2"
-            d="M50,24.7c-1.6,0-2.7-1.1-2.7-2.7v-2.7c0-1.6,1.1-2.7,2.7-2.7s2.7,1.1,2.7,2.7V22C52.7,23.6,51.6,24.7,50,24.7
-       z"
-          />
-          <path
-            class="st2"
-            d="M50,75.7c-1.6,0-2.7-1.1-2.7-2.7v-2.7c0-1.6,1.1-2.7,2.7-2.7s2.7,1.1,2.7,2.7V73C52.7,74.6,51.6,75.7,50,75.7
-       z"
-          />
-          <path
-            class="st2"
-            d="M76.9,48.9h-2.7c-1.6,0-2.7-1.1-2.7-2.7c0-1.6,1.1-2.7,2.7-2.7h2.7c1.6,0,2.7,1.1,2.7,2.7
-       C79.6,47.8,78.5,48.9,76.9,48.9z"
-          />
-          <path
-            class="st2"
-            d="M25.8,48.9h-2.7c-1.6,0-2.7-1.1-2.7-2.7c0-1.6,1.1-2.7,2.7-2.7h2.7c1.6,0,2.7,1.1,2.7,2.7
-       C28.5,47.8,27.4,48.9,25.8,48.9z"
-          />
-          <path
-            class="st2"
-            d="M68.8,30c-1.6,0-2.7-1.1-2.7-2.7c0-0.8,0.3-1.3,0.8-1.9l2.7-2.7c1.1-1.1,2.7-1.1,3.8,0c1.1,1.1,1.1,2.7,0,3.8
-       l-2.7,2.7C70.1,29.8,69.6,30,68.8,30z"
-          />
-          <path
-            class="st2"
-            d="M31.2,30c-0.8,0-1.3-0.3-1.9-0.8l-2.7-2.7c-1.1-1.1-1.1-2.7,0-3.8c1.1-1.1,2.7-1.1,3.8,0l2.7,2.7
-       c1.1,1.1,1.1,2.7,0,3.8C32.5,29.8,32,30,31.2,30z"
-          />
-          <path
-            class="st2"
-            d="M71.5,70.3c-0.8,0-1.3-0.3-1.9-0.8l-2.7-2.7c-1.1-1.1-1.1-2.7,0-3.8c1.1-1.1,2.7-1.1,3.8,0l2.7,2.7
-       c1.1,1.1,1.1,2.7,0,3.8C72.8,70.1,72.3,70.3,71.5,70.3z"
-          />
-          <path
-            class="st2"
-            d="M28.5,70.3c-1.6,0-2.7-1.1-2.7-2.7c0-0.8,0.3-1.3,0.8-1.9l2.7-2.7c1.1-1.1,2.7-1.1,3.8,0s1.1,2.7,0,3.8
-       l-2.7,2.7C29.9,70.1,29.3,70.3,28.5,70.3z"
-          />
-        </g>
-      </g>
+    <svg viewBox="0 0 24 24" width="30" height="30" fill="#FFC024">
+      <path d="M12.727 3.722A.736.736 0 0011.996 3a.734.734 0 00-.723.722v1.742c0 .39.332.721.723.721a.736.736 0 00.731-.721V3.722zm3.376 3.16a.735.735 0 000 1.02.726.726 0 001.031 0l1.239-1.236a.73.73 0 000-1.028.73.73 0 00-1.015 0l-1.255 1.244zm-9.245 1.02a.724.724 0 001.022 0c.283-.265.283-.746.009-1.02l-1.24-1.244a.748.748 0 00-1.022 0 .743.743 0 00-.008 1.02l1.239 1.244zm5.138-.132c-2.32 0-4.24 1.916-4.24 4.23 0 2.314 1.92 4.239 4.24 4.239 2.311 0 4.232-1.925 4.232-4.239s-1.92-4.23-4.232-4.23zm8.272 4.952c.4 0 .732-.332.732-.722a.736.736 0 00-.732-.722h-1.737a.734.734 0 00-.724.722c0 .39.333.722.724.722h1.737zM3.723 11.278A.734.734 0 003 12c0 .39.333.722.723.722h1.738c.399 0 .732-.332.732-.722a.736.736 0 00-.732-.722H3.723zm13.403 4.828a.74.74 0 00-1.023 0 .735.735 0 000 1.02l1.255 1.244a.73.73 0 001.015-.008.72.72 0 000-1.02l-1.247-1.236zM5.619 17.334a.743.743 0 00-.008 1.02.75.75 0 001.03.008l1.24-1.236a.727.727 0 00.008-1.02.75.75 0 00-1.031 0l-1.24 1.228zm7.108 1.202a.736.736 0 00-.731-.721.734.734 0 00-.723.721v1.742c0 .39.332.722.723.722a.736.736 0 00.731-.722v-1.742z"></path>
     </svg>
   );
 };
