+++ src/@portfolio-ui/ToggleSwitch.js
@@ -2,21 +2,21 @@ import React, { useContext } from 'react';
 import Switch from 'react-switch';
 import styled from 'styled-components';
 
+import { FormattedIcon } from '@portfolio-ui/';
+
 import ThemeToggleContext from '@context/ThemeToggleContext';
-import Sun from '@images/toggle/sun.png';
-import Moon from '@images/toggle/moon.png';
+
 
 const SwitchWrapper = styled(Switch)`
-  .sun {
-    height: 22px;
-    width: auto;
-    padding: 2px 0 0 8px;
+  svg {
+    height: 27px;
+    transform: rotate(-91.13deg);
   }
 
   .moon {
-    height: 20px;
-    width: auto;
-    padding: 2px 0 0 7px;
+    svg {
+    }
+    /* height: 20px; */
   }
 `;
 
@@ -31,8 +31,8 @@ export const ToggleSwitch = () => {
         checked={theme === 'dark' ? true : false}
         onColor="#555"
         offColor="#222"
-        checkedIcon={<img src={Moon} alt="moon icon" className="moon" />}
-        uncheckedIcon={<img src={Sun} alt="sun icon" className="sun" />}
+        checkedIcon={<FormattedIcon name="Moon" className="moon" />}
+        uncheckedIcon={<FormattedIcon name="Sun" />}
         boxShadow="0 0 2px 3px #2D85F4"
         activeBoxShadow="0 0 2px 3px #2D85F4"
         handleDiameter={23}
