+++ src/components/Seo/Seo.js
@@ -1,21 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
-import { useStaticQuery , graphql} from 'gatsby';
+import getSiteMetadata from '@lib/get-site-metadata.js';
 
 const SEO = ({ title, description, keywords, author }) => {
-  const data = useStaticQuery(graphql`
-    query {
-      site {
-        siteMetadata {
-          title
-          author
-          description
-          keywords
-        }
-      }
-    }
-  `);
+  const data = getSiteMetadata();
 
   const metaDescription = description || data.site.siteMetadata.description;
   const metaTitle = title || data.site.siteMetadata.title;
