+++ src/hooks/use-dark-mode.js
@@ -4,6 +4,8 @@ import useSWR from 'swr';
 export const themeStorageKey = 'adarsha:theme';
 
 const isServer = typeof window === 'undefined';
+
+// change value to dark if you want initial value to dark
 const getTheme = () => {
   if (isServer) return 'light';
   return localStorage.getItem(themeStorageKey) || 'light';
