+++ src/pages/blog.js
@@ -1,13 +1,13 @@
 import React from 'react';
 
 import Layout from '@components/Layout';
-import usePosts from '@hooks/use-posts';
+import getPosts from '@lib/get-posts';
 import SEO from '@components/Seo';
 import { SectionTitle } from '@portfolio-ui/';
 import BlogPreview from '@src/components/BlogPreview/';
 
 export default function Blog() {
-  const blogs = usePosts();
+  const blogs = getPosts();
 
   return (
     <Layout>
